work

+++ Code/Hawk/Util/Util_Console.cpp
@@ -6,8 +6,8 @@
 namespace Hawk {
 namespace Util_Console
 {
-	void SetRandSeed(unsigned int p_uiSeed);
-	void GetRand(float p_uiMin, float p_uiMax, unsigned int p_uiCount);
+	void SetRandSeed(UINT32 p_uiSeed);
+	void GetRand(float p_uiMin, float p_uiMax, UINT32 p_uiCount);
 }
 
 void Util_Console::Register(ConsoleCommandManager* p_pManager, std::shared_ptr<Dispatcher> p_Dispatcher)
@@ -16,20 +16,20 @@ void Util_Console::Register(ConsoleCommandManager* p_pManager, std::shared_ptr<D
 	p_pManager->Register("test.getRand", &Util_Console::GetRand, p_Dispatcher.get(), "Gets random numbers", "[min] [max] [count]");
 }
 
-void Util_Console::SetRandSeed(unsigned int p_uiSeed)
+void Util_Console::SetRandSeed(UINT32 p_uiSeed)
 {
 	Random::Initialize(p_uiSeed);
 }
 
-void Util_Console::GetRand(float p_uiMin, float p_uiMax, unsigned int p_uiCount)
+void Util_Console::GetRand(float p_uiMin, float p_uiMax, UINT32 p_uiCount)
 {
 	CONSOLE_WRITE_SCOPE();
-	static const unsigned int sc_uiMaxCols = 10;
-	unsigned int l_uiCurrCol = 0;
+	static const UINT32 sc_uiMaxCols = 10;
+	UINT32 l_uiCurrCol = 0;
 	std::cout << "\n";
-	for (unsigned int i = 0; i < p_uiCount; i++)
+	for (UINT32 i = 0; i < p_uiCount; i++)
 	{
-		std::cout << Random::GetFloat(p_uiMin, p_uiMax) << " ";
+		std::cout << Random::GetInt(p_uiMin, p_uiMax) << " ";
 
 		if (l_uiCurrCol == sc_uiMaxCols)
 		{
